requirements.txt
django-filter
django-allauth